# admin-panel/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Shortlinker API",
    "description": "Shortlinker management API",
    "contact": {
      "name": "AptS-1547",
      "email": "apts-1547@esaps.net"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.5.2"
  },
  "paths": {
...
    "/admin/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "登录验证 - 检查管理员 token",
        "operationId": "admin_login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginCredentials"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AuthSuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid administrator token"
          },
          "429": {
            "description": "Login rate limit exceeded"
          }
        }
      }
    },
...
    "/admin/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "刷新 token",
        "operationId": "refresh_admin_token",
        "responses": {
          "200": {
            "description": "Token refreshed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AuthSuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Refresh token is missing or invalid"
          },
          "429": {
            "description": "Refresh rate limit exceeded"
          }
        }
      }
    },
    "/admin/v1/auth/verify": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "验证 token - 如果中间件通过，则 token 有效",
        "operationId": "verify_admin_token",
        "responses": {
          "200": {
            "description": "Token is valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_MessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token is missing or invalid"
          }
        }
      }
    },
...
    "/admin/v1/config/{key}": {
      "get": {
        "tags": [
          "config"
        ],
        "summary": "获取单个配置",
        "operationId": "get_config",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "description": "Configuration key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Configuration item",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ConfigItemResponse"
                }
              }
            }
          },
          "404": {
            "description": "Configuration key not found"
          }
        }
      },
      "put": {
        "tags": [
          "config"
        ],
        "summary": "更新配置",
        "operationId": "update_config",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "description": "Configuration key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Configuration updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ConfigUpdateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid configuration value"
          },
          "404": {
            "description": "Configuration key not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_AuthSuccessResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "object",
            "description": "认证成功响应",
            "required": [
              "message",
              "expires_in"
            ],
            "properties": {
              "expires_in": {
                "type": "integer",
                "format": "int64",
                "minimum": 0
              },
              "message": {
                "type": "string"
              }
            }
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ApiResponse_ConfigItemResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "object",
            "description": "配置项响应",
            "required": [
              "key",
              "value",
              "value_type",
              "requires_restart",
              "is_sensitive",
              "updated_at"
            ],
            "properties": {
              "is_sensitive": {
                "type": "boolean"
              },
              "key": {
                "type": "string"
              },
              "requires_restart": {
                "type": "boolean"
              },
              "updated_at": {
                "type": "string"
              },
              "value": {
                "type": "string"
              },
              "value_type": {
                "$ref": "#/components/schemas/ValueType",
                "description": "值类型"
              }
            }
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ApiResponse_ConfigUpdateResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "object",
            "description": "配置更新响应",
            "required": [
              "key",
              "value",
              "requires_restart",
              "is_sensitive",
              "message"
            ],
            "properties": {
              "is_sensitive": {
                "type": "boolean"
              },
              "key": {
                "type": "string"
              },
              "message": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "requires_restart": {
                "type": "boolean"
              },
              "value": {
                "type": "string"
              }
            }
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ApiResponse_MessageResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "object",
            "description": "简单消息响应",
            "required": [
              "message"
            ],
            "properties": {
              "message": {
                "type": "string"
              }
            }
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ConfigUpdateRequest": {
        "type": "object",
        "description": "配置更新请求",
        "required": [
          "value"
        ],
        "properties": {
          "value": {
            "type": "string"
          }
        }
      },
...
      "LoginCredentials": {
        "type": "object",
        "required": [
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          }
        }
      },
...
      "ValueType": {
        "type": "string",
        "description": "配置值类型枚举\n\n用于标识配置项在数据库和前端的类型。",
        "enum": [
          "string",
          "int",
          "float",
          "bool",
          "json",
          "enum",
          "stringarray",
          "enumarray"
        ]
      }
    }
  },
...
}